Extract shared dashboard components
Move reusable dashboard badges, menus, editable titles, metadata rows, and regenerate controls out of Dashboard so later panel splits can focus on data flow.
